+++ Assets/Procedures/MaskToSdfTextureProcedure/Runtime/MaskToSdfTextureProcedure.cs
@@ -9,7 +9,7 @@ using UnityEngine.Rendering;
 
 namespace Simplex.Procedures
 {
-	public class ScalarTextureToSdfTextureProcedure
+	public class MaskToSdfTextureProcedure
 	{
 		RenderTexture _sdfTexture;
 		RenderTexture _floodTexture;
@@ -21,6 +21,7 @@ namespace Simplex.Procedures
 		int _ShowSeedsKernel;
 
 		LocalKeyword _ADD_BORDERS;
+		LocalKeyword[] _sourceChannelKeywords;
 
 		Vector2Int _groupThreadCount;
 
@@ -30,6 +31,7 @@ namespace Simplex.Procedures
 
 		[System.Serializable] public enum DownSampling { None, Half, Quater }
 		[System.Serializable] public enum Precision { _16, _32 }
+		[System.Serializable] public enum SourceChannel { R, G, B, Luminance }
 
 
 		static class ShaderIDs
@@ -41,13 +43,12 @@ namespace Simplex.Procedures
 			public static readonly int _Resolution = Shader.PropertyToID( nameof( _Resolution ) );
 			public static readonly int _StepSize = Shader.PropertyToID( nameof( _StepSize ) );
 			public static readonly int _SeedThreshold = Shader.PropertyToID( nameof( _SeedThreshold ) );
-			public static readonly int _TexelSize = Shader.PropertyToID( nameof( _TexelSize ) );
 		}
 
 
-		public ScalarTextureToSdfTextureProcedure()
+		public MaskToSdfTextureProcedure()
 		{
-			_computeShader = Object.Instantiate( Resources.Load<ComputeShader>( nameof( ScalarTextureToSdfTextureProcedure ) ) );
+			_computeShader = Object.Instantiate( Resources.Load<ComputeShader>( nameof( MaskToSdfTextureProcedure ) ) );
 			_computeShader.hideFlags = HideFlags.HideAndDontSave;
 
 			_SeedKernel = _computeShader.FindKernel( nameof( _SeedKernel ) );
@@ -56,13 +57,18 @@ namespace Simplex.Procedures
 			_ShowSeedsKernel = _computeShader.FindKernel( nameof( _ShowSeedsKernel ) );
 
 			_ADD_BORDERS = new LocalKeyword( _computeShader, nameof( _ADD_BORDERS ) );
+			var sourceChannelKeywordNames = System.Enum.GetNames( typeof( SourceChannel ) );
+			_sourceChannelKeywords = new LocalKeyword[ sourceChannelKeywordNames.Length ];
+			for( int sc = 0; sc < _sourceChannelKeywords.Length; sc++ ){
+				_sourceChannelKeywords[ sc ] = new LocalKeyword( _computeShader, "_" + sourceChannelKeywordNames[ sc ].ToUpper() );
+			}
 		}
 
 
 		public void Update
 		(
 			Texture sourceTexture, float sourceValueThreshold, 
-			DownSampling downSampling = DownSampling.None, Precision precision = Precision._32, bool addBorders = false,
+			SourceChannel sourceChannel = SourceChannel.R, DownSampling downSampling = DownSampling.None, Precision precision = Precision._32, bool addBorders = false,
 			bool _showSource = false
 		){
 
@@ -93,7 +99,6 @@ namespace Simplex.Procedures
 				_computeShader.SetTexture( _DistKernel, ShaderIDs._FloodTexRead, _floodTexture );
 				_computeShader.SetTexture( _ShowSeedsKernel, ShaderIDs._FloodTexRead, _floodTexture );
 				_computeShader.SetInts( ShaderIDs._Resolution, new int[]{ resolution.x, resolution.y } );
-				_computeShader.SetVector( ShaderIDs._TexelSize, _sdfTexture.texelSize );
 				_groupThreadCount = new Vector2Int(
 					Mathf.CeilToInt( resolution.x / (float) threadGroupWidth ),
 					Mathf.CeilToInt( resolution.y / (float) threadGroupWidth )
@@ -102,6 +107,12 @@ namespace Simplex.Procedures
 
 			// Set keywords.
 			if( _computeShader.IsKeywordEnabled( _ADD_BORDERS ) != addBorders ) _computeShader.SetKeyword( _ADD_BORDERS, addBorders );
+			int sourceChannelIndex = (int) sourceChannel;
+			for( int sc = 0; sc < _sourceChannelKeywords.Length; sc++ ){
+				if( _computeShader.IsKeywordEnabled( _sourceChannelKeywords[ sc ] ) != ( sc == sourceChannelIndex ) ){
+					_computeShader.SetKeyword( _sourceChannelKeywords[ sc ], sc == sourceChannelIndex );
+				}
+			}
 
 			// Seed.
 			_computeShader.SetTexture( _SeedKernel,  ShaderIDs._SeedTexRead, sourceTexture );
@@ -139,10 +150,11 @@ namespace Simplex.Procedures
 
 		static RenderTexture CreateTexture( string name, Vector2Int resolution, GraphicsFormat format )
 		{
-			RenderTexture rt = new RenderTexture( resolution.x, resolution.y, 0, format, 0 );
-			rt.name = name;
-			rt.autoGenerateMips = false;
-			rt.enableRandomWrite = true;
+			var rt = new RenderTexture( resolution.x, resolution.y, 0, format, 0 ){
+				name = name,
+				autoGenerateMips = false,
+				enableRandomWrite = true,
+			};
 			rt.Create();
 			return rt;
 		}